refactor: 서버 전용 유틸리티를 lib/server 폴더로 분리
- lib/server/ 폴더 생성하여 서버 전용 코드 분리
- posts.ts, mdx.ts를 lib/server/로 이동 (fs 사용)
- lib/utils/의 posts.ts, mdx.ts는 클라이언트/서버 겸용 유틸로 유지
- 불필요한 blog.ts, mdx-utils.ts 파일 정리
- 관련 import 경로 업데이트
